requirements.txt
wpilib >= 2018.0.9, < 2019.0
robotpy-ctre >= 2018.1.8, < 2019.0
robotpy-wpilib-utilities >= 2018.1.0, < 2019.0
pyfrc >= 2018.0.3, < 2019.0
numpy ~= 1.14
coverage
